server/agent/requirements.txt
anthropic>=0.42.0
websockets>=12.0
httpx>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pytest-asyncio>=0.23.0
